Avoid incorrect cast in sql_js in wasm builds
In order to avoid:

    Type 'double' is not a subtype of type 'int' in type cast

Raised when drift is built with `--wasm` it's needed to first cast
the value to double then turn it into int and it does makes
sense as int isn't native type in JS while double is the default
numeric type in JS.

+++ drift/lib/src/web/sql_js.dart
@@ -114,7 +114,8 @@ class SqlJsDatabase {
 
   /// Returns the `user_version` pragma from sqlite.
   int get userVersion {
-    return _selectSingleRowAndColumn('PRAGMA user_version;') as int;
+    return (_selectSingleRowAndColumn('PRAGMA user_version;') as double)
+        .toInt();
   }
 
   /// Sets sqlite's `user_version` pragma to the specified [version].
@@ -152,7 +153,8 @@ class SqlJsDatabase {
   /// [export].
   int lastInsertId() {
     // load insert id. Will return [{columns: [...], values: [[id]]}]
-    return _selectSingleRowAndColumn('SELECT last_insert_rowid();') as int;
+    return (_selectSingleRowAndColumn('SELECT last_insert_rowid();') as double)
+        .toInt();
   }
 
   dynamic _selectSingleRowAndColumn(String sql) {
